Rename debugoutput.go to gl_arb_debug_output.go
